+++ src/models/_entities/projects.rs
@@ -9,13 +9,13 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub pid: Uuid,
     pub name: String,
     #[sea_orm(column_type = "Text", nullable)]
     pub description: Option<String>,
     pub status: String,
-    pub created_by: i32,
+    pub created_by: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
